+++ RELEASE_NOTES.md
@@ -2,10 +2,12 @@
 
 New features:
  - Deployment of `idd` as an npm package. 
+ - Heatmap supports log10-based color palette. 
  - Palette editor `InteractiveDataDisplay.ColorPaletteEditor`.
  - IDD Knockout Js extended:
       - New binding [`iddAxisSettings`](https://github.com/predictionmachines/InteractiveDataDisplay/wiki/KnockoutJS%3A-Axis) allows creating a labelled axis, change tick font size etc.
       - New binding `iddPlotTitles` to provide titles for each of a plot properties.      
+      - New binding `iddEditorColorPalette` to assign a color palette to the color palette editor instance.
  - Box-and-whiskers plot (a kind of markers plot) allows to specify line thickness.
  - Plot has new property `padding` which allows to specify padding which is added to the bounding box of the plot in fit-to-view.
  - Bar plot allows to specify orientation, either horizontal or vertical.